requirements.txt
aiohttp==3.7.4.post0
anyio==3.5.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
multidict==6.0.2
rfc3986==1.5.0
sniffio==1.2.0
typing_extensions==4.2.0
yarl==1.7.2
